srr-nrcs/requirements.txt
overpy==0.6
pandas==2.1.0
PyYAML==6.0.1
geopy==2.4.1
requests==2.31.0
geopandas==0.13.2
fiona==1.9.4
matplotlib==3.8.0
contextily==1.5.0
pyproj==3.6.1
shapely==2.0.2 